+++ packages/picasso-codemod/src/v18.0.0/picasso-lab/picasso-lab.ts
@@ -33,11 +33,16 @@ import {
   ImportSpecifier,
   Transform,
   ImportDeclaration,
-  JSCodeshift,
-  ASTPath
+  JSCodeshift
 } from 'jscodeshift'
 
 import { replaceWith, isImportByPath, hasTopLevelComment } from '../../utils'
+import {
+  warnUsersAbout,
+  getSpecifierName,
+  getUnsolvableImportDeclarations,
+  getUnsolvableIdentifierNames
+} from './utils'
 
 type ImportTypes =
   | ImportSpecifier
@@ -97,20 +102,6 @@ const insertImports = (
   }
 }
 
-const getSpecifierName = (specifier: any) => {
-  const name = specifier.local!.name
-
-  if (specifier.type === 'ImportSpecifier') {
-    const identifierName = specifier.imported.loc!.identifierName
-
-    return identifierName && identifierName === name
-      ? name
-      : `${identifierName} as ${name}`
-  }
-
-  return name
-}
-
 const getImportsMap = (
   unsolvableIdentifierNames: Set<string>,
   filter: (node: ImportDeclaration) => boolean,
@@ -168,69 +159,6 @@ const removeNonNamespaceImports = (
       path => path.value.specifiers![0].type !== 'ImportNamespaceSpecifier'
     )
     .remove()
-}
-
-const getUnsolvableImportDeclarations = (
-  target: string,
-  root: ReturnType<Core>,
-  j: JSCodeshift
-) => {
-  const unsolvableImportDeclarations: ASTPath<ImportDeclaration>[] = []
-
-  root
-    .find(j.ImportDeclaration, ({ source }) => source.value === target)
-    .forEach(path => {
-      // we don't know what user would like to do with comments, thus, only user
-      // can decide what to do with imports that have comments.
-      if (path.value.comments) {
-        unsolvableImportDeclarations.push(path)
-      }
-    })
-
-  return unsolvableImportDeclarations
-}
-
-const warnUsersAbout = (
-  unsolvableImportDeclarations: ASTPath<ImportDeclaration>[],
-  filePath: string
-) => {
-  const totalImports = unsolvableImportDeclarations.length
-
-  if (totalImports > 0) {
-    console.log(
-      '\x1b[33m%s\x1b[0m',
-      `[***] Unresolved import statement${totalImports > 1 ? 's' : ''}:`
-    )
-  }
-
-  unsolvableImportDeclarations.forEach(elem => {
-    if (elem.value.loc?.start) {
-      const { line, column } = elem.value.loc?.start
-
-      console.log(`${filePath}:${line}:${column}`)
-    }
-  })
-
-  if (totalImports > 0) {
-    console.log(
-      '\x1b[33m%s\x1b[0m',
-      '[***] There are some unresolved import statements, please fix them.'
-    )
-  }
-}
-
-const getUnsolvableIdentifierNames = (
-  unsolvableImportDeclarations: ASTPath<ImportDeclaration>[]
-) => {
-  const unsolvableIdentifierNames: Set<string> = new Set()
-
-  unsolvableImportDeclarations.forEach(path => {
-    path.value.specifiers?.forEach(specifier =>
-      unsolvableIdentifierNames.add(getSpecifierName(specifier))
-    )
-  })
-
-  return unsolvableIdentifierNames
 }
 
 const transform: Transform = (file, api) => {
